Re-implement income estimation, handle minor benefits
